Numero total de vendas por cliente

+++ PDV/src/main/java/br/unipar/programacaointernet/pdv/dto/VendaDto.java
@@ -13,5 +13,5 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 public class VendaDto {
     private String nomeCliente;
-    private BigDecimal totalVendas;
+    private int totalVendas;
 }
